refactor: split html renderer support bands
Replace the mixed HTML renderer root with explicit highlighting, message-shaping, and template support bands, and remove the internal MarkdownRenderer name collision.
